requirements.txt
Django==2.1.4
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
pytz==2018.7
